+++ cligen/argcvt.nim
@@ -50,6 +50,8 @@ type ArgcvtParams* = object ## \
   shortNoVal*: set[char]  ## short options keys where value may be omitted
   longNoVal*: seq[string] ## long option keys where value may be omitted
   minStrQuoting*: bool    ## only quote string defaults when necessary
+  trueDefault*: string    ## How to render a default value of "true"
+  falseDefault*: string   ## How to render a default value of "false"
 
 proc argKeys*(a: ArgcvtParams, argSep="="): string =
   ## `argKeys` generates the option keys column in help tables
@@ -76,7 +78,8 @@ proc argParse*(dst: var bool, dfl: bool, a: var ArgcvtParams): bool =
   return true
 
 proc argHelp*(dfl: bool; a: var ArgcvtParams): seq[string] =
-  result = @[ a.argKeys(argSep=""), "bool", $dfl ]
+  result = @[ a.argKeys(argSep=""), "bool",
+              if dfl: a.trueDefault else: a.falseDefault ]
   if a.parSh.len > 0:
     a.shortNoVal.incl(a.parSh[0]) # bool can elide option arguments.
   a.longNoVal.add(move(a.parNm))  # So, add to *NoVal.